중간 통합본
