+++ front/react/package.json
@@ -14,12 +14,17 @@
     "@babel/core": "^7.9.0",
     "@babel/preset-env": "^7.9.5",
     "@babel/preset-react": "^7.9.4",
+    "@fortawesome/fontawesome-svg-core": "^1.2.28",
+    "@fortawesome/free-solid-svg-icons": "^5.13.0",
+    "@fortawesome/react-fontawesome": "^0.1.9",
     "axios": "^0.19.2",
     "babel-loader": "^8.1.0",
     "css-loader": "^3.5.3",
     "react": "^16.13.1",
+    "react-bootstrap": "^1.0.1",
     "react-dom": "^16.13.1",
     "react-search": "^2.1.1",
+    "react-stomp": "^5.0.0",
     "remarkable": "^2.0.0",
     "style-loader": "^1.2.1",
     "webpack": "^4.43.0",
@@ -28,7 +33,8 @@
   },
   "dependencies": {
     "bootstrap": "^4.4.1",
-    "react-bootstrap": "^1.0.1",
-    "react-router-dom": "^5.2.0"
+    "nav": "0.0.2",
+    "react-router-dom": "^5.2.0",
+    "table": "^5.4.6"
   }
 }
