+++ front/react/public/index.html
@@ -27,6 +27,11 @@
     type="text/css"
     href="http://localhost:8090/assets/footer.css"
   />
+  <link
+    rel="stylesheet"
+    type="text/css"
+    href="http://localhost:8090/assets/contents.css"
+  />
   <meta http-equiv="content-type" content="text/html; charset=utf-8" />
   <meta
     name="viewport"
