+++ src/main/java/com/douzone/codingvirus19/config/WebSecurityConfig.java
@@ -27,7 +27,7 @@ public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
 	@Override
 	protected void configure(HttpSecurity http) throws Exception {
 		http.csrf().disable().cors();
-		http.authorizeRequests().antMatchers("/assets/**","/api/join").permitAll()
+		http.authorizeRequests().antMatchers("/assets/**","/api/**","/chat/**").permitAll()
 //				.antMatchers("/auth/admin/**").hasRole("ADMIN")
 //				.antMatchers("/auth/**").hasAnyRole("ADMIN", "USER") // 내부적으로 접두어 "ROLE_"가 붙는다.
 				.anyRequest().authenticated();
@@ -35,7 +35,7 @@ public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
 		http.formLogin().loginPage("/") // default
 				.loginProcessingUrl("/user/auth")
 				.failureUrl("/error") // default
-				.defaultSuccessUrl("/main", true) // 로그인 성공시
+				.defaultSuccessUrl("/codingvirus19/main", true) // 로그인 성공시
 				.usernameParameter("email").passwordParameter("password").successHandler(authenticationSuccessHandler())
 				.permitAll();
