--- src/main/java/com/douzone/codingvirus19/controller/WebController.java
@@ -1,20 +1,0 @@
-package com.douzone.codingvirus19.controller;
-
-import org.springframework.boot.web.servlet.error.ErrorController;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class WebController implements ErrorController {
-	
-	@GetMapping({"/","/error","/main"})
-	public String index() {
-		return "main/index";
-	}
-	
-	@Override
-	public String getErrorPath() {
-		return "/error";
-	}
-	
-}
